Refactored about page
Summary: I've adjusted text. This will still need improvement ofcourse but I think it's a really good start. Please check the copy and if you have suggestions let me know.

Test Plan: Please read the text and if you have suggestions your welcome to give them to me. But I think it's still a litle early to give to much comments on it.

Reviewers: maarten

Reviewed By: maarten

Differential Revision: http://phab.wmc2.nl/D29